+++ core/src/main/java/net/ivpn/core/vpn/controller/WireGuardBehavior.java
@@ -93,6 +93,7 @@ public class WireGuardBehavior extends VpnBehavior implements ServiceConstants, 
         _fastestServer = server;
     };
     private LiveData<Server> fastestServer;
+    private long pauseDuration = 0;
 
     @Inject
     WireGuardBehavior(WireGuardKeyController wireGuardKeyController,
@@ -353,6 +354,7 @@ public class WireGuardBehavior extends VpnBehavior implements ServiceConstants, 
     }
 
     private void pauseVpn(long pauseDuration) {
+        this.pauseDuration = pauseDuration;
         LOGGER.info("pauseVpn: state = " + state);
         behaviourListener.onDisconnectingFromVpn();
         setState(PAUSING);
@@ -365,6 +367,8 @@ public class WireGuardBehavior extends VpnBehavior implements ServiceConstants, 
         LOGGER.info("Stop, state = " + state);
         timer.stopTimer();
         stopVpn();
+        setState(NOT_CONNECTED);
+        updateNotification();
     }
 
     private void stopVpn() {
@@ -401,7 +405,7 @@ public class WireGuardBehavior extends VpnBehavior implements ServiceConstants, 
     }
 
     private void updateNotification() {
-        updateNotification(0);
+        updateNotification(pauseDuration);
     }
 
     private void updateNotification(long pauseDuration) {